factor implementation out to plain objects
